requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Cache & Task Queue
redis==5.0.1
celery==5.3.6
flower==2.0.1

# HTTP Client
httpx==0.25.2

# Async Support
anyio==3.7.1

# Environment Variables
python-dotenv==1.0.0

# Validation
email-validator==2.1.0

# Rate Limiting
slowapi==0.1.9
slowapi==0.1.9  
